login and register done

+++ core/views.py
@@ -1,7 +1,12 @@
-from django.shortcuts import render,redirect
+from django.shortcuts import render,redirect,HttpResponse
+from django.contrib import messages
 from rest_framework import response ,viewsets,status
 from rest_framework.permissions import IsAuthenticated
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.decorators import login_required
 from .models import Task,Room,Topics
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
 from .serializers import TaskSerializer
 from .forms import RoomForm
 from django.db.models import Q
@@ -19,10 +24,51 @@ def home(request):
     context = {'rooms':rooms,'topics':topics,'room_count':room_count}
     return render(request,'home.html',context)
 
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')                
+        try:
+            user = User.objects.get(username=username)            
+        except:
+            messages.error(request,'User does not exist')
+        
+        user =authenticate(request,username=username,password=password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'Username or Password in invalid..')
+    context ={'page':page}     
+    return render(request,'loginRegister.html',context)
+
+def registerUser(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error as occured')   
+    return render(request,'loginRegister.html',{'form':form})
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
 def room (request,pk):
     room = Room.objects.get(id=pk)
     return render(request,'room.html',{'room':room})
 
+@login_required
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -33,10 +79,12 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'room_form.html',context)
 
-
+@login_required
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed to do this..")
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -45,8 +93,12 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request, 'room_form.html',context)
 
+@login_required
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed to do this..")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
